FIX: compiler warning
git-svn-id: https://www.nitrc.org/svn/cmtk/trunk@1049 42a5c34f-2066-0410-bec5-ba365beb4995

+++ core/libs/Registration/cmtkImagePairSimilarityJointHistogram.h
@@ -100,7 +100,7 @@ public:
    */
   template<class T> void Increment( const T a, const T b )
   {
-    this->m_JointHistogram.Increment( a, std::max<size_t>( 0, std::min<size_t>( this->m_NumberOfBinsY-1, b ) ) );
+    this->m_JointHistogram.Increment( static_cast<size_t>( a ), std::max<size_t>( 0, std::min<size_t>( this->m_NumberOfBinsY-1, static_cast<size_t>( b ) ) );
   }
 
   /// Add another metric object to this one.
